+++ bundle/src/assembly/resources/scenarios/FmuTest/application/EtsiApplication.json
@@ -1,0 +1,4 @@
+{
+    "minInterval": "500ms",
+    "maxInterval": "2s"
+}
