+++ bundle/src/assembly/resources/scenarios/FmuTest/scenario_config.json
@@ -1,0 +1,32 @@
+{
+    "simulation": {
+        "id": "HelloWorld",
+        "duration": "1000s",
+        "randomSeed": 268965854,
+        "projection": {
+            "centerCoordinates": {
+                "latitude": 52.23,
+                "longitude": 11.82
+            },
+            "cartesianOffset": {
+                "x": -691174.08,
+                "y": -5789894.65
+            }
+        },
+        "network": {
+            "netMask": "255.255.0.0",
+            "vehicleNet": "10.1.0.0",
+            "rsuNet": "10.2.0.0",
+            "tlNet": "10.3.0.0",
+            "csNet": "10.4.0.0",
+            "serverNet": "10.5.0.0",
+            "tmcNet": "10.6.0.0"
+        }
+    },
+    "federates": {
+        "application": true,
+        "sns": true,
+        "output": true,
+        "sumo": true
+    }
+}
