+++ bundle/src/assembly/resources/scenarios/FmuTest/sumo/highway.sumocfg
@@ -1,0 +1,10 @@
+<configuration>
+	<input>
+		<net-file value="highway.net.xml" />
+		<route-files value="highway.rou.xml" />
+	</input>
+	<time>
+		<begin value="0" />
+		<end value="10000" />
+	</time>
+</configuration>
